Move code from Main into Controller
