+++ src/zoo/data/penData/allPensData.json
@@ -1,1 +1,1 @@
-[{"area":200,"name":"Dry Pen 1","length":10,"width":20,"temp":17,"type":"DRY","penId":0,"assignedKeepers":[{"name":"Hardip","pensTrainedFor":["DRY","AVIARY"],"assignedPenIds":[]}],"animalIDsInPen":[0,7]},{"area":180,"name":"Petting Pen 1","length":12,"width":15,"temp":14,"type":"PETTING","penId":1,"assignedKeepers":[{"name":"Alan","pensTrainedFor":["DRY","PETTING"],"assignedPenIds":[]}],"animalIDsInPen":[2,3]},{"height":6,"waterVolume":300,"landArea":50,"name":"Amphibious Pen 1","length":10,"width":10,"temp":14,"type":"PARTDRYWATER","penId":2,"assignedKeepers":[{"name":"Alex","pensTrainedFor":["AQUARIUM","PARTDRYWATER"],"assignedPenIds":[]},{"name":"Alex","pensTrainedFor":["AQUARIUM","PARTDRYWATER"],"assignedPenIds":[2]}],"animalIDsInPen":[1,6]},{"height":6,"volume":720,"name":"Aquarium 1","length":12,"width":10,"temp":13,"type":"AQUARIUM","penId":3,"assignedKeepers":[{"name":"Farhad","pensTrainedFor":["AVIARY","AQUARIUM"],"assignedPenIds":[]}],"animalIDsInPen":[5]},{"height":7,"volume":1260,"name":"Aviary 1","length":12,"width":15,"temp":12,"type":"AVIARY","penId":4,"assignedKeepers":[{"name":"Hardip","pensTrainedFor":["DRY","AVIARY"],"assignedPenIds":[0]}],"animalIDsInPen":[4]}]
+[{"area":200,"name":"Dry Pen 1","length":10,"width":20,"temp":17,"type":"DRY","penId":0,"assignedKeepers":[{"name":"Hardip","pensTrainedFor":["DRY","AVIARY"],"assignedPenIds":[]},{"name":"Alan","pensTrainedFor":["DRY","PETTING"],"assignedPenIds":[1,0]}],"animalIDsInPen":[0,7]},{"area":180,"name":"Petting Pen 1","length":12,"width":15,"temp":14,"type":"PETTING","penId":1,"assignedKeepers":[{"name":"Alan","pensTrainedFor":["DRY","PETTING"],"assignedPenIds":[]}],"animalIDsInPen":[2,3]},{"height":6,"waterVolume":300,"landArea":50,"name":"Amphibious Pen 1","length":10,"width":10,"temp":14,"type":"PARTDRYWATER","penId":2,"assignedKeepers":[{"name":"Alex","pensTrainedFor":["AQUARIUM","PARTDRYWATER"],"assignedPenIds":[2]}],"animalIDsInPen":[1,6]},{"height":6,"volume":720,"name":"Aquarium 1","length":12,"width":10,"temp":13,"type":"AQUARIUM","penId":3,"assignedKeepers":[{"name":"Farhad","pensTrainedFor":["AVIARY","AQUARIUM"],"assignedPenIds":[]}],"animalIDsInPen":[5]},{"height":7,"volume":1260,"name":"Aviary 1","length":12,"width":15,"temp":12,"type":"AVIARY","penId":4,"assignedKeepers":[{"name":"Hardip","pensTrainedFor":["DRY","AVIARY"],"assignedPenIds":[0]}],"animalIDsInPen":[4]}]
