+++ src/zoo/data/penData/dryPensData.json
@@ -1,25 +1,1 @@
-[
-  {
-    "area": 200,
-    "name": "Dry Pen 1",
-    "length": 10,
-    "width": 20,
-    "temp": 17,
-    "type": "DRY",
-    "penId": 0,
-    "assignedKeepers": [
-      {
-        "name": "Hardip",
-        "pensTrainedFor": [
-          "DRY",
-          "AVIARY"
-        ],
-        "assignedPenIds": []
-      }
-    ],
-    "animalIDsInPen": [
-      0,
-      7
-    ]
-  }
-]
+[{"area":200,"name":"Dry Pen 1","length":10,"width":20,"temp":17,"type":"DRY","penId":0,"assignedKeepers":[{"name":"Hardip","pensTrainedFor":["DRY","AVIARY"],"assignedPenIds":[]},{"name":"Alan","pensTrainedFor":["DRY","PETTING"],"assignedPenIds":[1,0]}],"animalIDsInPen":[0,7]}]
