+++ src/zoo/data/penData/partDryWaterPensData.json
@@ -1,27 +1,1 @@
-[
-  {
-    "height": 6,
-    "waterVolume": 300,
-    "landArea": 50,
-    "name": "Amphibious Pen 1",
-    "length": 10,
-    "width": 10,
-    "temp": 14,
-    "type": "PARTDRYWATER",
-    "penId": 2,
-    "assignedKeepers": [
-      {
-        "name": "Alex",
-        "pensTrainedFor": [
-          "AQUARIUM",
-          "PARTDRYWATER"
-        ],
-        "assignedPenIds": [2]
-      }
-    ],
-    "animalIDsInPen": [
-      1,
-      6
-    ]
-  }
-]
+[{"height":6,"waterVolume":300,"landArea":50,"name":"Amphibious Pen 1","length":10,"width":10,"temp":14,"type":"PARTDRYWATER","penId":2,"assignedKeepers":[{"name":"Alex","pensTrainedFor":["AQUARIUM","PARTDRYWATER"],"assignedPenIds":[2]}],"animalIDsInPen":[1,6]}]
