+++ src/zoo/data/zooKeeperData/keeperData.json
@@ -1,43 +1,1 @@
-[
-  {
-    "name": "Hardip",
-    "pensTrainedFor": [
-      "DRY",
-      "AVIARY"
-    ],
-    "assignedPenIds": [
-      0,
-      4
-    ]
-  },
-  {
-    "name": "Alex",
-    "pensTrainedFor": [
-      "AQUARIUM",
-      "PARTDRYWATER"
-    ],
-    "assignedPenIds": [
-      2
-    ]
-  },
-  {
-    "name": "Farhad",
-    "pensTrainedFor": [
-      "AVIARY",
-      "AQUARIUM"
-    ],
-    "assignedPenIds": [
-      3
-    ]
-  },
-  {
-    "name": "Alan",
-    "pensTrainedFor": [
-      "DRY",
-      "PETTING"
-    ],
-    "assignedPenIds": [
-      1
-    ]
-  }
-]
+[{"name":"Hardip","pensTrainedFor":["DRY","AVIARY"],"assignedPenIds":[0,4]},{"name":"Alex","pensTrainedFor":["AQUARIUM","PARTDRYWATER"],"assignedPenIds":[2]},{"name":"Farhad","pensTrainedFor":["AVIARY","AQUARIUM"],"assignedPenIds":[3]},{"name":"Alan","pensTrainedFor":["DRY","PETTING"],"assignedPenIds":[1,0]}]
